Support inheriting metadata from directories.
This change means that tags, disabled and prefs can be set in specific
directories, or in specific test paths, as well as in specific tests.
In order to set it in directories create a /path/to_dir/__dir__.ini
file. These apply to each directory and all subdirectories.

Occasionally one wishes to override the metadata set in per-directory ini file
e.g. in order to enable one specific test whilst all others are disabled, or
to unset the prefs specified in higher-level directories. To achieve this the
concept of atoms has been introduced to the parser, which are used to represent
constant values. In particular:

disabled: @False

Will ensure that a test is enabled, whilst

prefs: [@Reset]

will ensure that no prefs from higher-level directories are used.